requirements.txt
# requirements.txt - M&A�X�N���C�s���O�c�[���ˑ��֌W

# �񓯊�HTTP����
aiohttp>=3.9.0,<4.0.0

# HTML���
beautifulsoup4>=4.12.0,<5.0.0
lxml>=6.0.0,<7.0.0

# Google Sheets�A�g
gspread>=5.12.0,<6.0.0
google-auth>=2.23.0,<3.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
google-auth-httplib2>=0.2.0,<1.0.0

# �u���E�U������
selenium>=4.15.0,<5.0.0
webdriver-manager>=4.0.0,<5.0.0

# �ݒ�t�@�C��
PyYAML>=6.0.1,<7.0.0

# �e�X�g�֘A
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0

# �J���E�f�o�b�O�p�i�I�v�V�����j
ipython>=8.12.0,<9.0.0
black>=23.9.0,<24.0.0
flake8>=6.1.0,<7.0.0